Move to a module in worker
